Add `repl` module
